Make default directory not relative to cwd
Before this change, the directory was always set to 'migrations' in the default case. This makes it dependent on the current working directory of any command being ran. The change uses current_app.root_path to make the default directory relative to the flask app it is for.

Additionally it uses _get_config for the init command, the utility function already used for the other functions (no behaviour change).