use unittest testrunner
... instead the deprecated `python setup.py tests`.

The latter one issued the following warning:

"WARNING: Testing via this command is deprecated and will be removed in
a future version. Users looking for a generic test entry point
independent of test runner are encouraged to use tox."

Another plus is that both the unittest runner and pytest also show
deprecation warnings, which `python setup.py tests` does not.