Fix running tests in venv with --system-site-packages
Ensure that flask is run via `python -m flask` rather than `flask`.
This is necessary to fix running the test suite when flask_migrate
is installed into a venv created with `--system-site-packages` and flask
itself is installed into the system site-packages directory.  This is
the scenario we are using in Gentoo to test packages prior to them
being installed.  Calling flask directly is a problem because it forces
`/usr/bin/pythonX.Y` and therefore omits the venv `sys.path` logic.